Refact: Rename 'LatestStoriesControls' => 'LatestStoriesInspectorControls'
